agents/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
openai
psycopg2-binary
python-dotenv
